contents reply sql read.jsp 권한수정

+++ src/main/webapp/WEB-INF/views/contents/read.jsp
@@ -316,8 +316,8 @@
 	          msg += "<span style='font-weight: bold;'>" + row.id + "</span>";
 	          msg += "  " + row.rdate;
 	          
-	          if ('${sessionScope.memberno}' == row.memberno || '${sessionScope.grade}' < 10) { // 글쓴이 일치여부 확인, 본인의 글만 삭제 가능함 ★
-	            msg += " <A href='javascript:reply_delete("+row.commentno+")'><IMG src='/contents/images/delete.png'></A>";
+	          if (($('#memberno', frm_reply).val().length != 0) && ('${sessionScope.memberno}' == row.memberno || '${sessionScope.grade}' < 10)) { // 글쓴이 일치여부 확인, 본인의 글만 삭제 가능함 ★	            
+		          msg += " <A href='javascript:reply_delete("+row.commentno+")'><IMG src='/contents/images/delete.png'></A>";
 	          }
 	          msg += "  " + "<br>";
 	          msg += row.content;
